backend
Signed-off-by: gbenhaim <gbenhaim@redhat.com>

impersonation

Signed-off-by: gbenhaim <gbenhaim@redhat.com>

proxy without ingress

Signed-off-by: gbenhaim <gbenhaim@redhat.com>

cleanup

Signed-off-by: gbenhaim <gbenhaim@redhat.com>

read token on boot

Signed-off-by: gbenhaim <gbenhaim@redhat.com>

add certificate for the proxy

Signed-off-by: gbenhaim <gbenhaim@redhat.com>

Remove UI ingress

We stopped using it in favor of a service with a node port.

Signed-off-by: gbenhaim <gbenhaim@redhat.com>

Long lived token and name fixes

Use long lived tokens for accessing the k8s api. This is needed
in order to simplify the generate of the nginx config with the token
that happens when the pod is created by an init container.

Also change some of the resources names to match their purpose.

Signed-off-by: gbenhaim <gbenhaim@redhat.com>

Add RBAC for users and demo users

Add the required RBAC file for a user to use konflux.
In addition create a test users and namespaces that use those.

Signed-off-by: gbenhaim <gbenhaim@redhat.com>

Local installation

Document local installation and make it possible by taking
out the inline scripts from the Github action.

Signed-off-by: gbenhaim <gbenhaim@redhat.com>

Deploy the certificate for the proxy

Signed-off-by: gbenhaim <gbenhaim@redhat.com>

Use pull secrets for the UI images

We temporary need to pull the UI images using a secret.

Signed-off-by: gbenhaim <gbenhaim@redhat.com>

fix path

Signed-off-by: gbenhaim <gbenhaim@redhat.com>

Comment out workspace-manager endpoints

Once the build of workspace-manager will be available
they can be added back.

Signed-off-by: gbenhaim <gbenhaim@redhat.com>

Debug the creation of the pull secret

Signed-off-by: gbenhaim <gbenhaim@redhat.com>