Deploy Keycloak
It will be used as the IDP. It's configured with 2 test users by defualt,
but it can also integrate with external IDPs such as Github, Google, etc.
The port nginx uses for https was changed to match the port which is
used to forward traffic from the host. It's required since oauth2-proxy
expects them to be the same (when it queries the details of the realm
from keycloak).

The endpoints are exposed through the nginx proxy.

The documentation used for creating this patch:

- https://www.keycloak.org/operator/basic-deployment
- https://www.keycloak.org/server/reverseproxy
https://stackoverflow.com/questions/66822016/where-are-all-of-the-keycloak-protocol-mapper-config-options-documented

Signed-off-by: gbenhaim <gbenhaim@redhat.com>