chore: handle errors on generate-err-logs.sh
The generate-err-logs is used to help us debug failures within the
cluster.
We encountered an issue where one of the namespaces failed the script
due to its pod missing the logs that are being read by the script.

Instead of failing, the script will skip the failure and log the
inability to access this resource.

In order to debug such situations better, added a print describing
which namespace's output is currently showing.

Signed-off-by: Omer Turner <oamsalem@redhat.com>