doc: strongly suggest increasing inotify resources
We mention that the user should increase the inotify resources only in
case they encounter failures due to too many open files.

We know that the user will most likely need it increased, and figuring
out the issue on a later stage might be challenging and frustrating.

To avoid that, we should recommend that the user will increase the limit
sooner rather than later.

Signed-off-by: Yftach Herzog <yherzog@redhat.com>