fix: undo adding of default service account
There was a bug in which the default service account was sometimes
created too late when testing creation of PVC. To address that, we're
now creating the SA explicitly.

Having said that, there's still seem to be a race condition between the
time the SA is created automatically and explicitly, which now causes
the opposite error: serviceaccounts "default" already exists.

Signed-off-by: Yftach Herzog <yherzog@redhat.com>