Consume image-controller from its repo
Pull the manifests required for the image-controller from the image-controller
repo. Update the image-controller image to match the version of the
manifests.

The ServiceMonitor isn't deployed since we don't deploy the Prometheous
operator. The metrics reader role isn't deployed as well for the same
reason.

Update the label selector of the image-controller pod in the deployment
script to match the label defined in the upstream manifest.

Signed-off-by: gbenhaim <gbenhaim@redhat.com>