Make impersonation optional
It's possible to deploy Konflux without using impersonation for the proxy
if the proxy and the cluster use the same IDP.

Allow this case by making impersonation optional.

In addition, fix the file permissions for the nginx conf which contains
a token. Make it readable only by the user and group.

Signed-off-by: Gal Ben Haim <gbenhaim@redhat.com>